requirements.txt
torch
torchvision
tqdm
numpy
opencv-python
pillow
